Fixed the extra zeroes being added to the end of the answer.

+++ HendrixJenna_Calculator/Main.h
@@ -35,8 +35,8 @@ public:
 
 	wxTextCtrl* answerBox;
 	
-	float num1;
-	float num2;
+	double num1;
+	double num2;
 	std::string number;
 	std::string mathFunction;
